+++ examples/regression/1-quickstart/plot_toy_model.py
@@ -2,7 +2,7 @@
 ======================================================
 Plotting MAPIE prediction intervals with a toy dataset
 ======================================================
-An example plot of :class:`mapie.regression.MapieRegressor` used
+An example plot of :class:`~mapie.regression.MapieRegressor` used
 in the Quickstart.
 """
 import numpy as np