+++ examples/regression/2-advanced-analysis/plot_both_uncertainties.py
@@ -2,8 +2,8 @@
 ================================================
 Estimating aleatoric and epistemic uncertainties
 ================================================
-This example uses :class:`mapie.regression.MapieRegressor` and
-:class:`mapie.quantile_regression.MapieQuantileRegressor` to estimate
+This example uses :class:`~mapie.regression.MapieRegressor` and
+:class:`~mapie.quantile_regression.MapieQuantileRegressor` to estimate
 prediction intervals capturing both aleatoric and epistemic uncertainties
 on a one-dimensional dataset with homoscedastic noise and normal sampling.
 """