+++ mapie/metrics.py
@@ -137,7 +137,7 @@ def regression_mean_width_score(
 def classification_mean_width_score(y_pred_set: ArrayLike) -> float:
     """
     Mean width of prediction set output by
-    :class:`mapie.classification.MapieClassifier`.
+    :class:`~mapie.classification.MapieClassifier`.
 
     Parameters
     ----------